Both in one file
The left one looks cut of because, um, it is. When you scroll down on
the page it’s suppose to show more of the cards like the three ones now
cut off. I didn’t draft that out though. The version on the right shows
to views, above and beyond the fold.